Update 1. js文件整理: 删除accounttool目录 ； 将tool、systemcontractv2目录中的公共js代码放入web3lib中
