--- accounttool/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "accounttool",
-  "version": "1.0.0",
-  "description": "tools for nodejs \"",
-  "main": "deploy.js",
-  "scripts": {
-    "test": "test"
-  },
-  "keywords": [
-    "tools"
-  ],
-  "author": "bcos",
-  "license": "GPL-3.0",
-  "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-preset-es2017": "^6.24.1",
-    "bignumber": "^1.1.0",
-    "bignumber.js": "^2.3.0",
-    "bn.js":"^4.11.6",
-    "crypto-js": "^3.1.6",
-    "express": "^4.15.3",
-    "keccak": "^1.2.0",
-    "rlp": "^2.0.0",
-    "secp256k1": "^3.2.5",
-    "web3": "^0.17.0-alpha"
-  },
-  "engines": {
-    "node": ">=6.10.3",
-    "npm": ">=3.10.0"
-  }
-}
