--- accounttool/readMe.md
@@ -1,10 +1,0 @@
-1. 依赖库
-	crypto
-	secp256k1
-	keccak
-	assert
-	rlp
-	bn.js
-
-
-
