+++ systemcontractv2/AuthorityManager.js
@@ -1,9 +1,9 @@
 var Web3 = require('web3');
-var config = require('./config');
+var config = require('../web3lib/config');
 var fs = require('fs');
 var BigNumber = require('bignumber.js');
-var web3sync = require('./web3sync');
-var sha3 = require("./sha3")
+var web3sync = require('../web3lib/web3sync');
+var sha3 = require("../web3lib/sha3")
 
 if (typeof web3 !== 'undefined') {
   web3 = new Web3(web3.currentProvider);
